Adds message logger as parent class to console and file logger.
