+++ h-logging/Configuration/IHMessageLogConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace Hylasoft.Logging.Configuration
+{
+  public interface IHMessageLogConfiguration : IHLoggingConfiguration
+  {
+    string TimestampFormat { get; }
+  }
+}
