+++ h-logging/Constants/ConfigurationDefaults.cs
@@ -2,19 +2,18 @@
 {
   public static class ConfigurationDefaults
   {
-    public static class File
-    {
-      public const string FileTargetName = "h-logging";
-
-      public const string LogName = "h-log";
-      
-      public const int MaxLogFiles = 7;
-    }
-
-    public static class Console
+    public static class Messages
     {
       public const string TimestampFormat = "HH:mm:ss";
 
+      public static class File
+      {
+        public const string FileTargetName = "h-logging";
+
+        public const string LogName = "h-log";
+
+        public const int MaxLogFiles = 7;
+      }
     }
   }
 }
