+++ h-logging/Loggers/HFileLogger.cs
@@ -1,18 +1,17 @@
 ﻿using System;
-using System.Linq;
 using Hylasoft.Logging.Configuration;
 using Hylasoft.Logging.Loggers.Base;
-using Hylasoft.Logging.Resources;
 using Hylasoft.Resolution;
 using NLog;
 using NLog.Config;
 using NLog.Targets;
+using OmniColour.Messages;
 
-using ConfigDefaults = Hylasoft.Logging.Constants.ConfigurationDefaults.File;
+using ConfigDefaults = Hylasoft.Logging.Constants.ConfigurationDefaults.Messages.File;
 
 namespace Hylasoft.Logging.Loggers
 {
-  public class HFileLogger : HLogger<IHFileLogConfiguration>
+  public class HFileLogger : HMessageLogger<IHFileLogConfiguration>
   {
     private readonly LogFactory _logFactory;
     private readonly LoggingConfiguration _logConfig;
@@ -27,30 +26,21 @@ namespace Hylasoft.Logging.Loggers
 
     protected string LogName { get { return _logName; } }
 
-    public HFileLogger(IHFileLogConfiguration config) : base(config)
+    public HFileLogger(IHFileLogConfiguration config)
+      : base(config)
     {
       _logName = ConfigDefaults.FileTargetName;
-      _logConfig = ToNlogConfiguration(config);
+      _logConfig = BuildNlogConfiguration();
       _logFactory = new LogFactory(LogConfig);
       _log = LogFactory.GetLogger(LogName);
     }
 
-    protected override Result LogIssues(ResultIssue[] issues)
-    {
-      return issues == null
-        ? Result.SingleWarning(Warnings.NothingToLog)
-        : issues.Where(issue => issue != null)
-          .Aggregate(Result.Success, (r, i) => r + LogIssue(i));
-    }
-
-    protected Result LogIssue(ResultIssue issue)
+    protected override Result LogMessage(IColourMessage message)
     {
       try
       {
-        var level = ToNlogLevel(issue.Level);
-        var message = BuildMessage(issue);
-
-        Log.Log(level, message);
+        var contents = message.BuildRaw();
+        Log.Log(LogLevel.Info, contents.Substring(0, contents.LastIndexOf(Environment.NewLine, StringComparison.Ordinal)));
         return Result.Success;
       }
       catch (Exception e)
@@ -64,32 +54,8 @@ namespace Hylasoft.Logging.Loggers
       return issue.Message;
     }
 
-    protected LogLevel ToNlogLevel(ResultIssueLevels level)
-    {
-      switch (level)
-      {
-        case ResultIssueLevels.Trace:
-          return LogLevel.Trace;
-
-        case ResultIssueLevels.Debug:
-          return LogLevel.Debug;
-
-        case ResultIssueLevels.Warning:
-          return LogLevel.Warn;
-
-        case ResultIssueLevels.Error:
-          return LogLevel.Error;
-
-        case ResultIssueLevels.Fatal:
-          return LogLevel.Fatal;
-
-        default:
-          return LogLevel.Info;
-      }
-    }
-
     #region Logging Configuration
-    private LoggingConfiguration ToNlogConfiguration(IHFileLogConfiguration config)
+    private LoggingConfiguration BuildNlogConfiguration()
     {
       var nlConfig = new LoggingConfiguration();
       var fileTarget = BuildFileTarget();
@@ -104,7 +70,7 @@ namespace Hylasoft.Logging.Loggers
     {
       const string fileFormat = "${{basedir}}/logs/{0}.log";
       const string layout = "${message}";
-      
+
       var fileName = string.Format(fileFormat, ReadConfig(c => c.LogName, ConfigDefaults.LogName));
       var maxFiles = ReadConfig(c => c.MaxLogFiles, ConfigDefaults.MaxLogFiles) ?? ConfigDefaults.MaxLogFiles;
