+++ h-logging/h-logging.csproj
@@ -57,9 +57,11 @@
     <Compile Include="Configuration\IHConsoleLogConfiguration.cs" />
     <Compile Include="Configuration\IHFileLogConfiguration.cs" />
     <Compile Include="Configuration\IHLoggingConfiguration.cs" />
+    <Compile Include="Configuration\IHMessageLogConfiguration.cs" />
     <Compile Include="Constants\ConfigurationDefaults.cs" />
     <Compile Include="IHLogger.cs" />
     <Compile Include="Loggers\Base\HLogger.cs" />
+    <Compile Include="Loggers\Base\HMessageLogger.cs" />
     <Compile Include="Loggers\HConsoleLogger.cs" />
     <Compile Include="Configuration\HLoggingDecorations.cs" />
     <Compile Include="Loggers\HFileLogger.cs" />
